+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/dot-edit-page-routing.module.ts
@@ -47,7 +47,7 @@ const dotEditPage: Routes = [
                 loadChildren: async () =>
                     (
                         await import(
-                            '../dot-experiments/experiments-shell/dot-experiments-shell.module'
+                            '../dot-experiments/dot-experiments-shell/dot-experiments-shell.module'
                         )
                     ).DotExperimentsShellModule
             }
