+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/index.ts
@@ -1,3 +1,3 @@
-export { DotExperimentsShellModule } from './experiments-shell/dot-experiments-shell.module';
+export { DotExperimentsShellModule } from './dot-experiments-shell/dot-experiments-shell.module';
 
 export { DotExperimentFeatureFlagResolver } from './shared/resolvers/dot-experiments-feature-flag-resolver.service';
